List adapter for classlist activity.
a list view is created for the classlist. so when each time a class is
selected, the exact class name will show on the top. I changed list
item from class to course so that the android studio won’t be confused.
Also, I commented out some fragment layout, and it goes to the activity
layouts now.